WoodenHorse vctr methods

Here are the methods listed for the WoodenHorse class. Refer to the documentation in the vctrs package to learn about their use.

Details

These are vctrs-methods that enable the WoodenHorse class, but should never be called directly by users. Consider these implementation details.

Functions

obj_print_header.WoodenHorse: See obj_print_header

vec_ptype_full.WoodenHorse: See vec_ptype_full

vec_ptype_abbr.WoodenHorse: See vec_ptype_abbr

vec_ptype2.BeechHorse: Generic for vec_ptype2

vec_ptype2.BeechHorse.BeechHorse: BeechHorse method for vec_ptype2.BeechHorse: attempt to merge GreekSoldier attributes.

vec_ptype2.OakHorse: Generic for vec_ptype2

vec_ptype2.OakHorse.OakHorse: OakHorse method for vec_ptype2.OakHorse: attempt to merge GreekSoldier attributes.

vec_cast.BeechHorse: Generic for vec_cast

vec_cast.BeechHorse.BeechHorse: BeechHorse method for vec_cast.BeechHorse: attempt to cast if the GreekSoldier attribute is similar.

vec_cast.OakHorse: Generic for vec_cast

vec_cast.OakHorse.OakHorse: OakHorse method for vec_cast.OakHorse: attempt to cast if the GreekSoldier attribute is similar.

vec_math.WoodenHorse: See vec_math. Exceptions to the math group generics for WoodenHorse class are is.na(), is.finite() and is.infinite().

vec_arith.WoodenHorse: Generic for vec_arith

vec_arith.WoodenHorse.default: Default method for vec_arith.WoodenHorse: try and apply arithmatic to the GreekSoldier attribute.

vec_arith.WoodenHorse.WoodenHorse: WoodenHorse method for vec_arith.WoodenHorse: try and apply arithmatic between the GreekSoldier attribute of the two vectors.

See also