ggnomics extensions to ggplot2

ggnomics relies on the extension mechanism of ggplot2 through ggproto class objects, which allows cross-package inheritance of objects such as geoms, stats, facets, scales and coordinate systems. These objects can be ignored by users for the purpose of making plots, since interacting with these objects is preferred through various geom_\*, stat_\*, facet_\*, coord_\* and scale_\* functions.

Note that the below says Functions, but actually describes ggproto classes.

Functions

ViewScaleS4: An child to ggplot's ViewScale ggproto that has a minor label getter. Note: this class is not exported.

ViewScaleS4Secondary: A child to ViewScaleS4 that has static break information. Note: this class is not exported.

CoordS4: See coord_S4

GeomRange: A child to GeomRects. See geom_range.

PositionDisjointRanges: See position_disjoint_ranges.

RangeS4: Identical to ggplot's Range ggproto. Note: this class and children are not exported and are for internal use.

RangeS4Continuous: A child to RangeS4 with S4 compatible train method.

RangeS4Discrete: A child to RangeS4 with S4 compatible train method.

ScaleS4: A child to ggplot's Scale ggproto. Note that this class is not exported but children are.

ScaleS4Continuous: A child to ScaleS4 and sibling to ggplot's ScaleContinuous.

ScaleS4Discrete: A child to ScaleS4 and sibling to ggplot's ScaleDiscrete

ScaleS4ContinuousPosition: A child to ScaleS4Continuous and sibling to ggplot's ScaleContinuousPosition. See scale_S4_continuous.

ScaleS4DiscretePosition: A child to ScaleS4Discrete and sibling to ggplot's ScaleDiscretePosition. See scale_S4_discrete.

ScaleGenomic: A child to ScaleS4ContinuousPosition for genomic coordinates. See scale_genomic.

See also