Axis for genomic positions

This axis guide is the genomic equivalent of guide_axis to pair with the genomic position scales scale_(x|y)_genomic().

guide_genomic_axis(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver()
)

Arguments

title
A character string or expression indicating a title of guide. If NULL, the title is not shown. By default (waiver()), the name of the scale object or the name specified in labs() is used for the title.

check.overlap
silently remove overlapping labels, (recursively) prioritizing the first, last, and middle labels.

angle
Compared to setting the angle in theme() / element_text(), this also uses some heuristics to automatically pick the hjust and vjust that you probably want.

n.dodge
The number of rows (for vertical axes) or columns (for horizontal axes) that should be used to render the labels. This is useful for displaying labels that would otherwise overlap.

order
Used to determine the order of the guides (left-to-right, top-to-bottom), if more than one guide must be drawn at the same location.

position
Where this guide should be drawn: one of top, bottom, left, or right.

Value

A guide object.

Details

This guide places the sequence names in the middle or their limits and places label for minor breaks at positions along the sequences. Defaults to guide_axis behaviour if not exposed to genomic data.

Methods (by generic)

guide_train: Trainer for genomic axis. See guide-exts.

guide_transform: Transformer for genomic axis. See guide-exts.

guide_gengrob: Graphic object generator for genomic axis. See guide-exts.

See also

Examples

NULL

#> NULL